+++ spec/nightwing/sidekiq/worker_stats_spec.rb
@@ -11,7 +11,7 @@ describe Nightwing::Sidekiq::WorkerStats do
         expect(subject.client).to receive(:increment).with("sidekiq.default.my_worker.processed")
         expect(subject.client).to receive(:increment).with("sidekiq.default.my_worker.finished")
 
-        subject.call(MyWorker.new, nil, "default") do
+        subject.call(MyWorker.new, {}, "default") do
           # beep
         end
       end
@@ -23,10 +23,22 @@ describe Nightwing::Sidekiq::WorkerStats do
         expect(subject.client).to receive(:increment).with("sidekiq.default.my_worker.failed")
 
         expect do
-          subject.call(MyWorker.new, nil, "default") do
+          subject.call(MyWorker.new, {}, "default") do
             fail "beep"
           end
         end.to raise_error RuntimeError
+      end
+    end
+
+    context "when being retried" do
+      it "increments process and retried count" do
+        expect(subject.client).to receive(:increment).with("sidekiq.default.my_worker.processed")
+        expect(subject.client).to receive(:increment).with("sidekiq.default.my_worker.retried")
+        expect(subject.client).to receive(:increment).with("sidekiq.default.my_worker.finished")
+
+        subject.call(MyWorker.new, { "retry" => 0, "retry_count" => 2 }, "default") do
+          # beep
+        end
       end
     end
   end